create useRoom hook
