+++ src/pages/Rooms.tsx
@@ -1,6 +1,7 @@
 import { useParams } from 'react-router';
 import { useAuth } from '../hooks/useAuth';
-import { FormEvent, useEffect, useState } from 'react';
+import { useRoom } from '../hooks/useRoom';
+import { FormEvent, useState } from 'react';
 import { Button } from '../components/Button';
 import { Question } from '../components/Question';
 import { RoomCode } from '../components/RoomCode';
@@ -8,62 +9,16 @@ import { database } from '../services/firebase';
 import logoImg from '../assets/images/logo.svg';
 import '../assets/styles/rooms.scss';
 
-type FirebaseQuestions = Record<string, {
-  author: {
-    name: string;
-    avatar: string;
-  }
-  content: string;
-  isAnswered: boolean;
-  isHighlighted: boolean;
-}>
-
 type RoomParams = {
   id: string;
-}
-
-type QuestionType = {
-  id: string;
-  author: {
-    name: string;
-    avatar: string;
-  }
-  content: string;
-  isAnswered: boolean;
-  isHighlighted: boolean;
 }
 
 export const Room = (): JSX.Element => {
   const { user } = useAuth();
   const params = useParams<RoomParams>();
   const roomId = params.id;
-  const [questions, setQuestions] = useState<QuestionType[]>([]);
-  const [title, setTitle] = useState([]);
-
-  useEffect(() => {
-    const roomRef = database.ref(`rooms/${roomId}`);
-
-    roomRef.on('value', room => {
-      const databaseRoom = room.val();
-      const firebaseQuestions: FirebaseQuestions = databaseRoom.questions ?? {};
-
-      const parsedQuestions = Object.entries(firebaseQuestions).map(([key, value]) => {
-        return {
-          id: key,
-          content: value.content,
-          author: value.author,
-          isHighlighted: value.isHighlighted,
-          isAnswered: value.isAnswered,
-        }
-      });
-
-      setTitle(databaseRoom.title);
-      setQuestions(parsedQuestions);
-    });
-
-  }, [roomId]);
-
   const [newQuestion, setNewQuestion] = useState('');
+  const { title, questions } = useRoom(roomId);
 
   const handleCreateSendQuestion = async (event: FormEvent) => {
     event.preventDefault();
